commit tree 051338a6be87db44c4970f7dd0b04102da8d1be9

+++ db/migrate/20210505235758_users.rb
@@ -4,7 +4,7 @@ class Users < ActiveRecord::Migration[5.2]
     create_table :users do |t|
       t.string :username
       t.string :password
-      t.integer :balance
+      t.decimal :balance
     end
   end
